+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/immutableArraysModule/SequenceExtensionsGenerator.kt
@@ -2,6 +2,7 @@ package com.danrusu.pods4k.immutableArrays.immutableArraysModule
 
 import com.danrusu.pods4k.immutableArrays.BaseType
 import com.danrusu.pods4k.immutableArrays.ImmutableArrayConfig
+import com.danrusu.pods4k.utils.controlFlow
 import com.danrusu.pods4k.utils.createFile
 import com.danrusu.pods4k.utils.function
 import com.danrusu.pods4k.utils.statement
@@ -31,8 +32,9 @@ private fun FileSpec.Builder.addSequenceToImmutableArray() {
             if (baseType == BaseType.GENERIC) {
                 addTypeVariable(baseType.type as TypeVariableName)
             }
-            statement("val elements = this.toList()")
-            statement("return ${baseType.generatedClassName}(elements.size) { elements[it] }")
+            controlFlow("return build${baseType.generatedClassName}") {
+                statement("addAll(this@toImmutableArray)")
+            }
         }
     }
 }